chore: Rename "ParentContainer" to "BaseContainer"
- Navbar에 useRouter 경로이름을 props로 받도록 수정
- 경로이름에 dashboard가 포함되어 있는 경우 Navbar의 utils 부분 보이도록
  함
